# api/openapi.yaml
openapi: "3.1.0"
info:
  title: Akashi API
  description: |
    Decision trace layer for multi-agent AI systems.

    Akashi records, queries, and audits every decision made by AI agents —
    providing "git blame for AI decisions." It captures the full decision
    lifecycle: what was decided, why, what alternatives were considered,
    and what evidence was used.

    ## Authentication

    All `/v1/` and `/billing/` endpoints require a Bearer JWT token obtained
    from `POST /auth/token`. Tokens are scoped to an organization and carry
    the agent's RBAC role.

    ## Rate Limits

    Rate limits are enforced per agent (or per IP for auth endpoints).
    Admin+ roles are exempt. When rate-limited, the API returns HTTP 429
    with error code `RATE_LIMITED`.

    | Category | Limit | Endpoints |
    |----------|-------|-----------|
    | Ingestion | 300/min | runs, events, trace |
    | Query | 300/min | query, temporal, history, recent, usage, conflicts, check |
    | Search | 100/min | search |
    | Auth | 20/min | token, signup |

    ## Response Envelope

    All JSON responses are wrapped in a standard envelope:

    ```json
    {
      "data": { ... },
      "meta": {
        "request_id": "uuid",
        "timestamp": "2026-01-15T10:30:00Z"
      }
    }
    ```

    Error responses use a parallel structure:

    ```json
    {
      "error": {
        "code": "INVALID_INPUT",
        "message": "description"
      },
      "meta": {
        "request_id": "uuid",
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/agents:
    post:
      operationId: createAgent
      tags: [Agents]
      summary: Create an agent
      description: |
        Register a new agent within the caller's organization.
        Requires `admin` role or higher.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Agent created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse_Agent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      operationId: listAgents
      tags: [Agents]
      summary: List all agents
      description: |
        List all agents in the caller's organization.
        Requires `admin` role or higher.
      responses:
        "200":
          description: List of agents.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse_AgentList"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /v1/agents/{agent_id}/tags:
    patch:
      operationId: updateAgentTags
      tags: [Agents]
      summary: Update an agent's tags
      description: |
        Replace the tags for an agent. Tags enable group-based access control:
        agents sharing at least one tag can see each other's decisions without
        explicit grants. Tags coexist with the existing grant system.
        Requires `admin` role or higher.
      parameters:
        - $ref: "#/components/parameters/AgentIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentTagsRequest"
      responses:
        "200":
          description: Tags updated. Returns the updated agent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse_Agent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Runs ───────────────────────────────────────────────────────────
...
  /v1/grants:
    post:
      operationId: createGrant
      tags: [Access]
      summary: Grant access to another agent
      description: |
        Create a fine-grained access grant allowing another agent to read
        or write specific resources. Agents can grant access to their own
        traces; admins can grant access to any resource.
        Requires `agent` role or higher.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGrantRequest"
      responses:
        "201":
          description: Grant created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse_AccessGrant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /v1/grants/{grant_id}:
    delete:
      operationId: deleteGrant
      tags: [Access]
      summary: Revoke an access grant
      description: |
        Delete an access grant. Agents can delete their own grants; admins
        can delete any grant. Requires `agent` role or higher.
      parameters:
        - name: grant_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the grant to delete.
      responses:
        "204":
          description: Grant deleted.
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Billing ────────────────────────────────────────────────────────
...
  /v1/usage:
    get:
      operationId: getUsage
      tags: [Billing]
      summary: Get organization usage and limits
      description: |
        Returns the current billing period's usage counters and plan limits
        for the authenticated agent's organization.
        Requires `reader` role or higher. Rate limited (query: 300/min).
      responses:
        "200":
          description: Usage and limit details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIResponse_UsageResponse"
        "429":
          $ref: "#/components/responses/RateLimited"

  # ── Export ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Ed25519-signed JWT obtained from `POST /auth/token`.
        Include as `Authorization: Bearer <token>`.
  parameters:
...
    AgentIDPath:
      name: agent_id
      in: path
      required: true
      schema:
        type: string
      description: The string identifier of the agent.
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
...
    Forbidden:
      description: Insufficient permissions.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    Conflict:
      description: Resource already exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    RateLimited:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
  schemas:
...
    ResponseMeta:
      type: object
      required: [request_id, timestamp]
      properties:
        request_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    APIError:
      type: object
      required: [error, meta]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - INVALID_INPUT
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - RATE_LIMITED
            - QUOTA_EXCEEDED
            - INTERNAL_ERROR
        message:
          type: string
        details:
          description: Additional context about the error.

    # ── Auth schemas ─────────────────────────────────────────────────
...
    AgentRole:
      type: string
      enum:
        - platform_admin
        - org_owner
        - admin
        - agent
        - reader
    Agent:
      type: object
      required: [id, agent_id, org_id, name, role, tags, metadata, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        agent_id:
          type: string
        org_id:
          type: string
          format: uuid
        name:
          type: string
        role:
          $ref: "#/components/schemas/AgentRole"
        tags:
          type: array
          items:
            type: string
          description: |
            Tags for group-based access control. Agents sharing at least one
            tag can see each other's decisions. Must match `^[a-z][a-z0-9_-]*$`.
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAgentRequest:
      type: object
      required: [agent_id, name, api_key]
      properties:
        agent_id:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/AgentRole"
        api_key:
          type: string
        tags:
          type: array
          items:
            type: string
          description: Optional tags for group-based access control.
        metadata:
          type: object
          additionalProperties: true
    UpdateAgentTagsRequest:
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          items:
            type: string
          description: |
            New set of tags for the agent. Replaces existing tags entirely.
            Each tag must match `^[a-z][a-z0-9_-]*$`.
...
    AccessGrant:
      type: object
      required: [id, org_id, grantor_id, grantee_id, resource_type, permission, granted_at]
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        grantor_id:
          type: string
          format: uuid
        grantee_id:
          type: string
          format: uuid
        resource_type:
          type: string
        resource_id:
          type: string
        permission:
          type: string
        granted_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
    CreateGrantRequest:
      type: object
      required: [grantee_agent_id, resource_type, permission]
      properties:
        grantee_agent_id:
          type: string
        resource_type:
          type: string
        resource_id:
          type: string
        permission:
          type: string
        expires_at:
          type: string
          description: RFC 3339 timestamp.

    # ── Run schemas ──────────────────────────────────────────────────
...
    UsageResponse:
      type: object
      required: [org_id, plan, period, decision_count, decision_limit, agent_limit]
      properties:
        org_id:
          type: string
          format: uuid
        plan:
          type: string
        period:
          type: string
        decision_count:
          type: integer
        decision_limit:
          type: integer
        agent_limit:
          type: integer
...
    APIResponse_Agent:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    APIResponse_AgentList:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    APIResponse_AccessGrant:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AccessGrant"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    APIResponse_UsageResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/UsageResponse"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...